Added protectedService utils and refactored validateRequestAgainstProtectedService to validateRequestAgainstWhitelistedUrls
