+++ lib/bindUtils/index.js
@@ -1,13 +1,17 @@
+const validateRequestAgainstWhitelistedUrls = require('../protectedService/validateRequestAgainstWhitelistedUrls.js');
+function setupProtectedServiceUtils () {
+	falafel.utils.protectedService = {
+		validateRequestAgainstWhitelistedUrls
+	};
+}
 
 const processBody = require('../rawHttpRequest/processBody');
 const validateUrlInput = require('../rawHttpRequest/validateUrlInput');
-const validateRequestAgainstProtectedService = require('../rawHttpRequest/validateRequestAgainstProtectedService.js');
 const formatOutput = require('../rawHttpRequest/formatOutput');
 function setupRawHttpRequestUtils () {
 	falafel.utils.rawHttpRequest = {
 		validateUrlInput,
 		processBody,
-		validateRequestAgainstProtectedService,
 		formatOutput
 	};
 }
