+++ lib/protectedService/validateRequestAgainstWhitelistedUrls.js
