+++ lib/rawHttpRequest/rawHttpRequestModel.js
@@ -2,7 +2,7 @@ const _ = require('lodash');
 
 const processBody = require('./processBody.js');
 const validateUrlInput = require('./validateUrlInput.js');
-const validateRequestAgainstProtectedService = require('./validateRequestAgainstProtectedService.js');
+const validateRequestAgainstWhitelistedUrls = require('../protectedService/validateRequestAgainstWhitelistedUrls.js');
 const formatOutput = require('./formatOutput.js');
 
 function convertArrayFormatToObject (targetArray) {
@@ -66,7 +66,7 @@ module.exports = {
 		Performs validation of the request against protected service whitelist
 		if enabled and configured
 	*/
-	beforeRequest: validateRequestAgainstProtectedService,
+	beforeRequest: validateRequestAgainstWhitelistedUrls,
 
 	//Transform the API response into a standardised format
 	afterSuccess: formatOutput
